backend/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
google-auth-oauthlib
google-auth-httplib2
google-api-python-client
pydantic-settings